chore: Update bodyParser.urlencoded configuration to not use extended mode
This commit updates the configuration of the `bodyParser.urlencoded` middleware in the `/routes/auth.js` file to not use the extended mode. By setting `extended` to `false`, the middleware will only parse key-value pairs in the URL-encoded data, improving security and performance. This change ensures that the application adheres to best practices for handling URL-encoded data.