refactor: Add support for uploading background pictures in /users/:userId route
This commit adds support for uploading background pictures in the /users/:userId route. It introduces a new variable `background_picture_id` to store the ID of the uploaded background picture. If a background picture is included in the request, it is saved to the database and associated with the user. The code also handles the file upload and moves the picture to the appropriate directory. This change enhances the user profile functionality by allowing users to customize their background pictures.