refactor: Add review picture location to SQL queries
This commit modifies the SQL queries in `sql_commands.js` to include the `review_picture_location` column from the `images` table. This allows retrieving the location of the review picture when fetching review cards. The added column is joined with the `reviews` table using a LEFT JOIN. This change enhances the functionality of the application by providing the review picture location along with other review details.