refactor: Fix user creation bug in auth route
This commit fixes a bug in the auth route where the user creation query was not properly executed due to a syntax error. The bug occurred when inserting a new user into the database, causing the application to throw an error. This fix ensures that the user creation query is executed correctly, allowing new users to be added to the database without any issues.