chore: Update .gitignore to include certificates
This commit updates the .gitignore file to include the "*.pem" files, which are certificates used by the application. By adding this entry to the .gitignore file, the certificates will be ignored by Git and will not be tracked in the repository. This change ensures that sensitive certificate files are not accidentally committed and pushed to the remote repository.