feat: Add Facebook authentication support
This commit adds support for Facebook authentication by including the passport-facebook npm dependency and configuring the FacebookStrategy in the auth.js file. The FacebookStrategy is used to authenticate users using their Facebook accounts. This feature allows users to log in to the application using their Facebook credentials, enhancing the authentication options available to users.