refactor: Update file upload limits in index.js
This commit updates the file upload limits in the index.js file. The previous limit allowed for only one file to be uploaded, but this update increases the limit to allow for two files to be uploaded. This enhancement improves the functionality of the application by allowing users to upload multiple files simultaneously.